made BlogCards.js its own module so that we can apply both the global and blog styles to the page at once.
